Q version 56.751: FIX BUG: pseudomacro ^NJ would not jump backwards
The code omitted to cast an unsigned quantity to signed before adding to the
macro position.

Updated:

 help/cn-j: Add an actual example of a backward jump, now it's fixed.

 q/q_version.h: 56.751

 q/scrdit.c: mcposn += (int16_t)thisch;

+++ q/scrdit.c
@@ -902,7 +902,7 @@ process_pseudo_arg(scrbuf5 *Curr, bool in_cmd)
   if (fornj)
   {
     fornj = false;
-    mcposn += thisch;              /* Do the jump */
+    mcposn += (int16_t)thisch;     /* Do the (signed!) jump */
     if (mcposn >= 0 && mcposn < scmacs[curmac]->maclen)
       GETNEXTCHR;
     err = "^NJ off macro end or start";
